Relocate register manipulation logic
