+++ AECompiler.Core/CodeGeneration/RegisterDescriptors/Register.cs
@@ -1,4 +1,4 @@
-﻿namespace AECompiler.Core.CodeGeneration
+﻿namespace AECompiler.Core.CodeGeneration.RegisterDescriptors
 {
     internal sealed class Register
     {
