+++ AECompiler.Core/CodeGeneration/RegisterDescriptors/RegisterName.cs
@@ -1,4 +1,4 @@
-﻿namespace AECompiler.Core.CodeGeneration
+﻿namespace AECompiler.Core.CodeGeneration.RegisterDescriptors
 {
     public enum RegisterName
     {
