+++ AECompiler.Core/CodeGeneration/RegisterDescriptors/State.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Collections.Generic;
+using AECompiler.Core.CodeGeneration.RegisterDescriptors;
 
 namespace AECompiler.Core.CodeGeneration
 {
